Add ByteRead/WriteChannel implementations for native files
Adds a native implementation of of ktor ByteReadChannel and
ByteWriteChannel which read/write to/from a file descriptor (to be
used with the ksrpc asConnection method). The implementations simply
create a thread dedicated to doing blocking reads/writes, and pipe
through a ByteChannel, so they are not the most efficient.